backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
faiss-cpu
pillow
python-multipart
numpy
scikit-learn
matplotlib
